feat: drop old-arch and update to more flexible and reusable and compact
