+++ components/shared/footer-component.tsx
