+++ docs/dashboard-case.png
